chore: split ast-utils into module per util
