+++ packages/css-selector-parser/src/index.ts
@@ -2,5 +2,6 @@ export { parseCssSelector } from "./selector-parser";
 export type { ParseConfig } from "./selector-parser";
 export * from "./ast-types";
 export { stringifySelectorAst } from "./stringify";
-export { walk, groupCompoundSelectors } from "./ast-utils";
-export type { WalkOptions } from "./ast-utils";
+export { walk } from "./ast-tools/walk";
+export type { WalkOptions } from "./ast-tools/walk";
+export { groupCompoundSelectors } from "./ast-tools/compound";
